requirements.txt
Django==3.2.11
djangorestframework==3.12.4
django-filter==22.1
requests==2.27.1
drf-yasg==1.21.4
gunicorn==20.1.0
python-dotenv==0.20.0
psycopg2-binary==2.8.6